# one-ui/ui/api/swagger/mocks.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample of a DataHub Swagger document defining endpoints to be mocked",
    "version": "1.0.0",
    "title": "Data Hub Curation API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:4200",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/search": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search documents",
        "description": "....",
        "operationId": "searchDocuments",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/AnyObject"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "search"
      },
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Get a document by URI",
        "description": "....",
        "operationId": "getDocumentByURI",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "docUri",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/AnyObject"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "search"
      }
    },
    "/facet-values/range": {
      "post": {
        "tags": [
          "facet values"
        ],
        "summary": "Get min and max range for facet value",
        "description": "....",
        "operationId": "getRangeForFacet",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/rangeForFacet"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "facetValue"
      }
    },
    "/facet-values": {
      "post": {
        "tags": [
          "facet values"
        ],
        "summary": "Get facet values that match a pattern",
        "description": "....",
        "operationId": "getFacetValues",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/facetValue"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/error"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        },
        "x-swagger-router-controller": "facetValue"
      }
    },
...
  },
  "definitions": {
    "AnyObject": {
      "type": "object",
      "description": "This is a temporary definition for any object until details can be fleshed out",
      "properties": {}
    },
...
    "rangeForFacet": {
      "type": "object",
      "description": "Range for facet value",
      "properties": {
        "referenceType": {
          "type": "string",
          "description": "element or field"
        },
        "entityTypeId": {
          "type": "string",
          "description": "identifier for the entity type in entity model"
        },
        "propertyPath": {
          "type": "string",
          "description": "path to property"
        }
      },
      "example": {
        "min": 1,
        "max": 31
      }
    },
    "facetValue": {
      "type": "object",
      "description": "Range for facet value",
      "properties": {
        "facetInfo": {
          "type": "object",
          "description": "facet information",
          "properties": {
            "referenceType": {
              "type": "string",
              "description": "element or field"
            },
            "entityTypeId": {
              "type": "string",
              "description": "identifier for the entity type in entity model"
            },
            "propertyPath": {
              "type": "string",
              "description": "path to property"
            }
          }
        },
        "limit": {
          "type": "integer",
          "description": "facet value limit"
        },
        "dataType": {
          "type": "string",
          "description": "facet data type"
        },
        "pattern": {
          "type": "string",
          "description": "pattern to search"
        }
      },
      "example": [
        "QUEDE",
        "QUEEN",
        "QUICK"
      ]
    },
...
    "error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  },
...
}